requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
ctranslate2==3.17.1
decorator==4.4.2
et-xmlfile==1.1.0
faster-whisper==0.7.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
lxml==4.9.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mutagen==1.46.0
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pgvector==0.2.0
Pillow==10.0.0
proglog==0.1.10
protobuf==4.23.4
psycopg2==2.9.6
pycryptodomex==3.18.0
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.19
sympy==1.12
tokenizers==0.13.3
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
websockets==11.0.3
yarl==1.9.2
yt-dlp==2023.7.6
